add variant list for users add variant list for admin

+++ app/src/main/java/com/example/user/votechain/Database/IVariantDataSource.java
@@ -7,9 +7,9 @@ import java.util.List;
 import io.reactivex.Flowable;
 
 public interface IVariantDataSource {
-    Flowable<List<VoteVariant>> getAllByVote(int userId);
+    Flowable<List<VoteVariant>> getAllByVote(long userId);
     void insert(VoteVariant... variants);
     void update(VoteVariant... variants);
     void  delete(VoteVariant variant);
-    void deleteAllByVote(int userId);
+    void deleteAllByVote(long userId);
 }
